requirements.txt
python-telegram-bot[all]==21.7
python-dotenv==1.0.1
aiohttp==3.10.10
qrcode==8.0
Flask==3.0.3
asgiref==3.8.1
Pillow==11.0.0
PyMuPDF==1.24.14
requests==2.32.3